+++ renderer/src/components/json-render/index.vue
@@ -1,0 +1,22 @@
+<template>
+    <el-scrollbar width="100%">
+        <div v-html="renderJson(json)">
+        </div>
+    </el-scrollbar>
+</template>
+
+<script setup lang="ts">
+import { computed, defineProps, PropType } from 'vue';
+import { renderJson } from '../main-panel/chat/markdown/markdown';
+
+const props = defineProps({
+    json: {
+        type: Object as PropType<string | object | undefined>,
+        required: true
+    }
+});
+
+</script>
+
+<style>
+</style>